+++ src/client/js/functions.js
@@ -1,7 +1,7 @@
 /*
  * pwix:modal/src/client/js/functions.js
  *
- * The public methods exposed through the pwixModal global object.
+ * The public methods exposed through the Modal global object.
  */
 
 import { mdModal } from '../classes/md_modal.class.js';
@@ -14,14 +14,14 @@ const _isJSObject = function( obj ){
 
 /**
  * @summary Enable/disable a button
- *  Only if a specific footer has not been set via pwixModal.setFooter()
+ *  Only if a specific footer has not been set via Modal.setFooter()
  * @locus Client
  * @param {String} button the button to enable/disable
  * @param {Boolean} enable whether the button should be enabled
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  */
-pwixModal.buttonEnable = function( button, enable, id ){
-    const btn = pwixModal.buttonFind( button, id );
+Modal.buttonEnable = function( button, enable, id ){
+    const btn = Modal.buttonFind( button, id );
     if( btn ){
         btn.prop( 'disabled', !enable );
     } else {
@@ -31,14 +31,14 @@ pwixModal.buttonEnable = function( button, enable, id ){
 
 /**
  * @summary Find a button element
- *  Only if a specific footer has not been set via pwixModal.setFooter()
+ *  Only if a specific footer has not been set via Modal.setFooter()
  * @locus Client
  * @param {String} button the button to search for
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  * @eturns {Object} the found button as a jQuery object, or null
  */
-pwixModal.buttonFind = function( button, id ){
-    const modal = pwixModal._client.Stack.modal( id );
+Modal.buttonFind = function( button, id ){
+    const modal = Modal._client.Stack.modal( id );
     if( modal ){
         return modal.buttonFind( button );
     }
@@ -48,8 +48,8 @@ pwixModal.buttonFind = function( button, id ){
  * @summary Close the topmost opened dialog
  * @locus Client
  */
-pwixModal.close = function(){
-    const modal = pwixModal._client.Stack.pop();
+Modal.close = function(){
+    const modal = Modal._client.Stack.pop();
     if( modal ){
         modal.close();
     }
@@ -59,15 +59,15 @@ pwixModal.close = function(){
  * @locus Client
  * @returns {Integer} the count of opened modals
  */
-pwixModal.count = function(){
-    return pwixModal._client.Stack.count();
+Modal.count = function(){
+    return Modal._client.Stack.count();
 };
 
 /**
  * @locus Client
  * @return {Array} known buttons
  */
-pwixModal.knownButtons = function(){
+Modal.knownButtons = function(){
     return [
         MD_BUTTON_OK,
         MD_BUTTON_CANCEL,
@@ -84,9 +84,9 @@ pwixModal.knownButtons = function(){
  * @param {Object} parms the running parameters of the new dialog
  * @returns {String} the identifier of this new modal
  */
-pwixModal.run = function( parms ){
+Modal.run = function( parms ){
     const modal = new mdModal( parms );
-    pwixModal._client.Stack.push( modal );
+    Modal._client.Stack.push( modal );
     return modal.id();
 };
 
@@ -96,8 +96,8 @@ pwixModal.run = function( parms ){
  * @param {String} template the name of the template
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  */
-pwixModal.setBody = function( template, id ){
-    const modal = pwixModal._client.Stack.modal( id );
+Modal.setBody = function( template, id ){
+    const modal = Modal._client.Stack.modal( id );
     if( modal ){
         modal.body( template );
     }
@@ -105,7 +105,7 @@ pwixModal.setBody = function( template, id ){
 
 /**
  * @summary Set the buttons of the currently opened dialog
- *  Only if a specific footer has not been set via pwixModal.setFooter()
+ *  Only if a specific footer has not been set via Modal.setFooter()
  * @locus Client
  * @param {Array|String} buttons the button or array of buttons to be set
  *  Only set the provided buttons if valid.
@@ -113,9 +113,9 @@ pwixModal.setBody = function( template, id ){
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  * @returns {Boolean} whether the provided buttons were valid and have been set
  */
-pwixModal.setButtons = function( buttons, id ){
+Modal.setButtons = function( buttons, id ){
     const btns = Array.isArray( buttons ) ? buttons : [ buttons ];
-    const knowns = pwixModal.knownButtons();
+    const knowns = Modal.knownButtons();
     let ok = true;
     btns.every(( btn ) => {
         if( !knowns.includes( btn )){
@@ -124,7 +124,7 @@ pwixModal.setButtons = function( buttons, id ){
         return ok;
     });
     if( ok ){
-        const modal = pwixModal._client.Stack.modal( id );
+        const modal = Modal._client.Stack.modal( id );
         if( modal ){
             modal.buttons( btns );
         }
@@ -140,8 +140,8 @@ pwixModal.setButtons = function( buttons, id ){
  * @param {String} classes the classes to be added to the '.modal' element
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  */
-pwixModal.setClasses = function( classes, id ){
-    const modal = pwixModal._client.Stack.modal( id );
+Modal.setClasses = function( classes, id ){
+    const modal = Modal._client.Stack.modal( id );
     if( modal ){
         modal.classes( classes );
     }
@@ -153,8 +153,8 @@ pwixModal.setClasses = function( classes, id ){
  * @param {String} template the name of the Blaze template
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  */
-pwixModal.setFooter = function( template, id ){
-    const modal = pwixModal._client.Stack.modal( id );
+Modal.setFooter = function( template, id ){
+    const modal = Modal._client.Stack.modal( id );
     if( modal ){
         modal.footer( template );
     }
@@ -165,9 +165,9 @@ pwixModal.setFooter = function( template, id ){
  *  OBSOLETED
  *  WILL BE REMOVED ON 2.0 VERSION
  */
-pwixModal.setTarget = function(){
+Modal.setTarget = function(){
     console.warn( 'pwix:modal setTarget() obsoleted method, redirected to target()' );
-    pwixModal.target( ...arguments );
+    Modal.target( ...arguments );
 };
 
 /**
@@ -176,8 +176,8 @@ pwixModal.setTarget = function(){
  * @param {String} title the title to be set
  * @param {String} id the identifier of the targeted dialog, defaulting to the topmost
  */
-pwixModal.setTitle = function( title, id ){
-    const modal = pwixModal._client.Stack.modal( id );
+Modal.setTitle = function( title, id ){
+    const modal = Modal._client.Stack.modal( id );
     if( modal ){
         modal.title( title );
     }
@@ -191,13 +191,13 @@ pwixModal.setTitle = function( title, id ){
  *  - id: a string identifier which identifies the modal, defaulting to the topmost one
  * @returns {jQuery} the target of the modal, either explicitely identified or the topmost one, or null
  */
-pwixModal.target = function( o ){
+Modal.target = function( o ){
     //console.debug( o );
     let modal;
     if( o ){
         if( _isJSObject( o )){
             if( o.target ){
-                modal = pwixModal._client.Stack.modal( o.id );
+                modal = Modal._client.Stack.modal( o.id );
                 if( modal ){
                     modal.target( o.target );
                 }
@@ -208,7 +208,7 @@ pwixModal.target = function( o ){
             console.error( 'pwix:modal target() expects an Object argument, found', o );
         }
     } else {
-        modal = pwixModal._client.Stack.modal();
+        modal = Modal._client.Stack.modal();
     }
     return modal ? modal.target() : null;
 };