+++ src/common/js/global.js
@@ -2,6 +2,6 @@
  * pwix:modal/src/common/js/global.js
  */
 
-pwixModal = {
+Modal = {
     _conf: {}
 };